Refactor code, fix bug.
