+++ src/main/java/org/bigtows/notebook/NotebookAccessible.java
@@ -5,7 +5,7 @@
  * Licensed under the MIT License. See LICENSE file in the project root for license information.
  */
 
-package org.bigtows.service.note.notebook;
+package org.bigtows.notebook;
 
 /**
  * Simple interface of credential
