+++ src/main/java/org/bigtows/notebook/Task.java
@@ -1,4 +1,4 @@
-package org.bigtows.service.note.notebook;
+package org.bigtows.notebook;
 
 public interface Task {
