+++ src/main/java/org/bigtows/notebook/evernote/EvernoteNotebookLoader.java
@@ -1,0 +1,4 @@
+package org.bigtows.notebook.evernote;
+
+public class EvernoteNotebookLoader {
+}
