+++ src/main/java/org/bigtows/notebook/evernote/creadential/ServiceType.java
@@ -5,7 +5,7 @@
  * Licensed under the MIT License. See LICENSE file in the project root for license information.
  */
 
-package org.bigtows.service.note.notebook.evernote.creadential;
+package org.bigtows.notebook.evernote.creadential;
 
 public enum ServiceType {
     SANDBOX,
