+++ src/main/java/org/bigtows/notebook/evernote/exception/SaveNotesException.java
@@ -1,4 +1,4 @@
-package org.bigtows.service.note.notebook.evernote.exception;
+package org.bigtows.notebook.evernote.exception;
 
 public class SaveNotesException extends RuntimeException {
