+++ src/main/java/org/bigtows/notebook/evernote/exception/StorageException.java
@@ -1,4 +1,4 @@
-package org.bigtows.service.note.notebook.evernote.exception;
+package org.bigtows.notebook.evernote.exception;
 
 public class StorageException extends RuntimeException {
