+++ src/main/java/org/bigtows/notebook/evernote/parser/TaskContext.java
@@ -5,12 +5,16 @@
  * Licensed under the MIT License. See LICENSE file in the project root for license information.
  */
 
-package org.bigtows.service.note.notebook.evernote.parser;
+package org.bigtows.notebook.evernote.parser;
+
+
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * Context for parsing XML Evernote note vision
  */
-public class SubTaskContext {
+public class TaskContext {
 
     /**
      * Unique identification
@@ -26,6 +30,12 @@ public class SubTaskContext {
      * Task name
      */
     private String name;
+
+    /**
+     * List sub task
+     */
+    private List<SubTaskContext> subTaskContexts = new ArrayList<>();
+
 
     /**
      * This task completed
@@ -63,6 +73,25 @@ public class SubTaskContext {
     public void setName(String name) {
         this.name = name;
     }
+
+    /**
+     * Gets sub task contexts.
+     *
+     * @return the sub task contexts
+     */
+    public List<SubTaskContext> getSubTaskContexts() {
+        return subTaskContexts;
+    }
+
+    /**
+     * Add sub task contexts.
+     *
+     * @param subTaskContexts the sub task contexts
+     */
+    public void addSubTaskContexts(SubTaskContext subTaskContexts) {
+        this.subTaskContexts.add(subTaskContexts);
+    }
+
 
     /**
      * Get Unique identification
