--- src/main/java/org/bigtows/service/note/Target.java
@@ -1,4 +1,0 @@
-package org.bigtows.service.note;
-
-public interface Target {
-}
