--- src/main/java/org/bigtows/service/note/notebook/evernote/EvernoteNotebookLoader.java
@@ -1,4 +1,0 @@
-package org.bigtows.service.note.notebook.evernote;
-
-public class EvernoteNotebookLoader {
-}
