+++ src/main/java/org/bigtows/utils/PinNoteIcon.java
@@ -10,6 +10,10 @@ import java.util.Optional;
 
 public class PinNoteIcon {
 
+    public static final Icon NOTIFICATION_PIN_NOTE = PinNoteIcon.getSvgIconFromResource(
+            "/icons/toolWindowRight_dark.svg", 1.2f
+    ).orElse(AllIcons.Ide.FatalError);
+
     public static final Icon TAB_EVERNOTE_ICON = PinNoteIcon.getSvgIconFromResource(
             "/icons/evernote.svg", 0.5f
     ).orElse(AllIcons.Ide.FatalError);
